+++ UnityUtils/Assets/Testing Scripts/TestingScript.cs
@@ -7,6 +7,7 @@ public class TestingScript : MonoBehaviour {
   public Vector3 testingVector3;
 
   public SFX sfx;
+  AudioClip clip;
   // Starter is called once before the first execution of Update after the MonoBehaviour is created
   void Start() {
     ObjectAnimations.AnimateTransformRotation(transform, new Vector3(transform.localRotation.x, transform.localRotation.y, transform.localRotation.z), new Vector3(4, 50, 90), 2);
@@ -14,5 +15,7 @@ public class TestingScript : MonoBehaviour {
     ObjectAnimations.AnimateTransformPosition(transform, transform.position, new Vector3(0, 5, 0), 2, animationCurve: curve);
 
     ObjectAnimations.AnimateSpriteRendererOpacity(GetComponent<SpriteRenderer>(), 0.5f, 1, 2);
+
+    SFX test = new SFX(clip).SetPitch(3, 4);
   }
 }