+++ UnityUtils/Packages/com.mydimons.unityutils/Runtime/Script Utils/Objects/ObjectModifiers/ModifierData.cs
@@ -1,5 +1,5 @@
-﻿namespace UnityUtils.ScriptUtils.Objects {
-  public class ObjectModifierData<T> {
+﻿namespace UnityUtils.ScriptUtils.Objects.Modifiers {
+  public class ModifierData<T> {
     /// <summary>
     /// Specifies the classType of modifier applied to the object.
     /// </summary>
@@ -11,11 +11,11 @@
     public T modifierValue;
 
     /// <summary>
-    /// Initializes a new Instance of the ObjectModifierData class with the <see cref="ObjectModifiers.ModifierType"/> and value.
+    /// Initializes a new Instance of the <see cref="ModifierData{T}"/> class with the <see cref="ModifierManager.ModifierType"/> and value.
     /// </summary>
     /// <param name="modifierType">The classType of modifier</param>
     /// <param name="modifierValue">The value associated with the modifier</param>
-    public ObjectModifierData(ModifierType modifierType, T modifierValue) {
+    public ModifierData(ModifierType modifierType, T modifierValue) {
       this.modifierType = modifierType;
       this.modifierValue = modifierValue;
     }